pull from netmq

+++ src/NetMQ/Core/SocketBase.cs
@@ -183,6 +183,8 @@ namespace NetMQ.Core
                     return new XSub(parent, threadId, socketId);
                 case ZmqSocketType.Stream:
                     return new Stream(parent, threadId, socketId);
+                case ZmqSocketType.Peer:
+                    return new Peer(parent, threadId, socketId);
                 default:
                     throw new InvalidException("SocketBase.Create called with invalid type of " + type);
             }
@@ -417,10 +419,8 @@ namespace NetMQ.Core
             // Process pending commands, if any.
             ProcessCommands(0, false);
 
-            string protocol;
-            string address;
 
-            DecodeAddress(addr, out address, out protocol);
+            DecodeAddress(addr, out string address, out string protocol);
 
             CheckProtocol(protocol);
 
@@ -541,9 +541,7 @@ namespace NetMQ.Core
         /// <exception cref="FaultException">the socket bind failed</exception>
         public int BindRandomPort([NotNull] string addr)
         {
-            string address, protocol;
-
-            DecodeAddress(addr, out address, out protocol);
+            DecodeAddress(addr, out string address, out string protocol);
 
             if (protocol != Address.TcpProtocol)
                 throw new ProtocolNotSupportedException("Address must use the TCP protocol.");
@@ -573,9 +571,7 @@ namespace NetMQ.Core
             // Process pending commands, if any.
             ProcessCommands(0, false);
 
-            string address;
-            string protocol;
-            DecodeAddress(addr, out address, out protocol);
+            DecodeAddress(addr, out string address, out string protocol);
 
             CheckProtocol(protocol);
 
@@ -702,7 +698,7 @@ namespace NetMQ.Core
             bool icanhasall = protocol == Address.PgmProtocol || protocol == Address.EpgmProtocol;
             Pipe newPipe = null;
 
-            if (!m_options.DelayAttachOnConnect || icanhasall)
+            if (!m_options.DelayAttachOnConnect || icanhasall || m_options.SocketType == ZmqSocketType.Peer)
             {
                 // Create a bi-directional pipe.
                 ZObject[] parents = { this, session };
@@ -771,10 +767,8 @@ namespace NetMQ.Core
             //  (from launch_child() for example) we're asked to terminate now.
             ProcessCommands(0, false);
 
-            string protocol;
-            string address;
 
-            DecodeAddress(addr, out address, out protocol);
+            DecodeAddress(addr, out string address, out string protocol);
 
             CheckProtocol(protocol);
 
@@ -783,8 +777,7 @@ namespace NetMQ.Core
                 if (UnregisterEndpoint(addr, this))
                     return;
 
-                Pipe pipe;
-                if (!m_inprocs.TryGetValue(addr, out pipe))
+                if (!m_inprocs.TryGetValue(addr, out Pipe pipe))
                     throw new EndpointNotFoundException("Endpoint was not found and cannot be disconnected");
 
                 pipe.Terminate(true);
@@ -792,8 +785,7 @@ namespace NetMQ.Core
             }
             else
             {
-                Endpoint endpoint;
-                if (!m_endpoints.TryGetValue(addr, out endpoint))
+                if (!m_endpoints.TryGetValue(addr, out Endpoint endpoint))
                     throw new EndpointNotFoundException("Endpoint was not found and cannot be disconnected");
 
                 endpoint.Pipe?.Terminate(false);
@@ -1209,7 +1201,7 @@ namespace NetMQ.Core
 
         /// <summary>
         /// Handle input-ready events by receiving and processing any incoming commands.
-        /// </summary>        
+        /// </summary>
         public virtual void InEvent()
         {
             // This function is invoked only once the socket is running in the context
@@ -1294,7 +1286,7 @@ namespace NetMQ.Core
 
         public void Hiccuped(Pipe pipe)
         {
-            if (m_options.DelayAttachOnConnect)
+            if (m_options.DelayAttachOnConnect && m_options.SocketType != ZmqSocketType.Peer)
                 pipe.Terminate(false);
             else
                 // Notify derived sockets of the hiccup
@@ -1356,9 +1348,7 @@ namespace NetMQ.Core
                 return;
             }
 
-            string address;
-            string protocol;
-            DecodeAddress(addr, out address, out protocol);
+            DecodeAddress(addr, out string address, out string protocol);
 
             CheckProtocol(protocol);
 
@@ -1543,6 +1533,8 @@ namespace NetMQ.Core
                     return "PULL";
                 case ZmqSocketType.Push:
                     return "PUSH";
+                case ZmqSocketType.Peer:
+                    return "PEER";
                 default:
                     return "UNKNOWN";
             }
